fix: cache ajv schemas to prevent memory leak (#47)
* fix: cache ajv schemas to prevent memory leak

There was an issue with the validation methods causing memory leaks, as
the schemas were being redefined on each run. Calling "ajv.compile" on
this methods was causing this memory leak, as they generated a new
schema everytime, even when called with an schema we had used earlier.
So, the more times we called one of these methods, the more duplicated
schemas we'd have compiled and stored in memory.

In order to fix this issue, "ajv.compile" has been replaced by
"ajv.addSchema" and "ajv.getSchema". The first is being used to compile
the schema, but it will only be run once: if the schema already existed,
we will reuse the former instance rather than recompiling over again.

* chore: upgrade ajv to latest version

* refactor: call getSchema only once if schema exists

* fix: error with urls as schema names

* fix: dependencies

* docs: add note on slashes not supported in schema names

* refactor: rename schema name suffix for unique param

Co-authored-by: kevinccbsg <kevinccbsg@gmail.com>